pyproject.toml
[tool.poetry]
name = "MusicBot2"
version = "0.0.0"
description = "A discord music bot written in Python."
authors = ["xNykram"]

[tool.poetry.dependencies]
python = "3.11.6"
youtube_search = "2.1.2"
pydantic = "2.4.2"
yt-dlp = "2023.10.13"
pynacl = "1.5.0"
'discord.py' = "2.3.2"
requests = "2.31.0"
beautifulsoup4 = "4.12.2"
sqlalchemy = "2.0.22"
pydantic-settings = "2.0.3"
psycopg2-binary = "*"

[tool.poetry.dev-dependencies]
debugpy = "1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'

[tool.isort]
atomic = true
profile = "black"
line_length = 88
skip_gitignore = true

[tool.pylint.messages-control]
disable = [
    "C0114","C0115","C0116","E1101","W0613","W0622"
]
